feat : 좋아요,댓글,대댓글,팔로우 시 알림 등록

See merge request s07-webmobile1-sub1/S07P11C208!117

+++ backend/src/main/java/com/gwangjubob/livealone/backend/service/impl/DealServiceImpl.java
@@ -1,19 +1,14 @@
 package com.gwangjubob.livealone.backend.service.impl;
 
 
-import com.gwangjubob.livealone.backend.domain.entity.DealCommentEntity;
-import com.gwangjubob.livealone.backend.domain.entity.DealEntity;
-import com.gwangjubob.livealone.backend.domain.entity.UserEntity;
-import com.gwangjubob.livealone.backend.domain.entity.UserLikeDealsEntity;
-import com.gwangjubob.livealone.backend.domain.repository.DealCommentRepository;
-import com.gwangjubob.livealone.backend.domain.repository.DealRepository;
-import com.gwangjubob.livealone.backend.domain.repository.UserLikeDealsRepository;
-import com.gwangjubob.livealone.backend.domain.repository.UserRepository;
+import com.gwangjubob.livealone.backend.domain.entity.*;
+import com.gwangjubob.livealone.backend.domain.repository.*;
 import com.gwangjubob.livealone.backend.dto.Deal.DealCommentDto;
 import com.gwangjubob.livealone.backend.dto.Deal.DealDto;
 import com.gwangjubob.livealone.backend.mapper.DealCommentMapper;
 import com.gwangjubob.livealone.backend.mapper.DealMapper;
 import com.gwangjubob.livealone.backend.service.DealService;
+import org.aspectj.weaver.ast.Not;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -29,15 +24,18 @@ public class DealServiceImpl implements DealService {
 
     private DealCommentRepository dealCommentRepository;
     private UserLikeDealsRepository userLikeDealsRepository;
+    private NoticeRepository noticeRepository;
     private DealCommentMapper dealCommentMapper;
     @Autowired
-    DealServiceImpl(DealRepository dealRepository, DealMapper dealMapper, UserRepository userRepository, DealCommentRepository dealCommentRepository, DealCommentMapper dealCommentMapper, UserLikeDealsRepository userLikeDealsRepository){
+    DealServiceImpl(DealRepository dealRepository, DealMapper dealMapper, UserRepository userRepository, DealCommentRepository dealCommentRepository,
+                    NoticeRepository noticeRepository, DealCommentMapper dealCommentMapper, UserLikeDealsRepository userLikeDealsRepository){
         this.dealRepository = dealRepository;
         this.dealMapper = dealMapper;
         this.userRepository = userRepository;
         this.dealCommentRepository = dealCommentRepository;
         this.dealCommentMapper = dealCommentMapper;
         this.userLikeDealsRepository = userLikeDealsRepository;
+        this.noticeRepository = noticeRepository;
     }
 
 
@@ -133,6 +131,33 @@ public class DealServiceImpl implements DealService {
             data.setUserNickname(dealComment.getUser().getNickname());
             data.setUserId(dealComment.getUser().getId());
             data.setPostIdx(dealComment.getDeal().getIdx());
+
+            if(!deal.getUser().getId().equals(user.getId())){
+                if (dealComment.getUpIdx() == 0) {
+                    NoticeEntity notice = NoticeEntity.builder()
+                            .noticeType("comment")
+                            .user(deal.getUser())
+                            .fromUserId(user.getId())
+                            .postType("deal")
+                            .postIdx(deal.getIdx())
+                            .commentIdx(dealComment.getIdx())
+                            .build();
+
+                    noticeRepository.save(notice);
+                }else{
+                    NoticeEntity notice = NoticeEntity.builder()
+                            .noticeType("reply")
+                            .user(deal.getUser())
+                            .fromUserId(user.getId())
+                            .postType("deal")
+                            .postIdx(deal.getIdx())
+                            .commentIdx(dealComment.getIdx())
+                            .commentUpIdx(dealComment.getUpIdx())
+                            .build();
+
+                    noticeRepository.save(notice);
+                }
+            }
         } else{
             data = null;
         }
@@ -156,11 +181,49 @@ public class DealServiceImpl implements DealService {
     }
 
     @Override
-    public boolean deleteDealComment(Integer idx) {
+    public boolean deleteDealComment(Integer idx, String userId) {
         Optional<DealCommentEntity> optionalDealComment = dealCommentRepository.findById(idx);
         if(optionalDealComment.isPresent()){
             DealCommentEntity dealComment = optionalDealComment.get();
-            dealCommentRepository.delete(dealComment);
+            DealEntity deal = dealRepository.findById(dealComment.getDeal().getIdx()).get();
+            UserEntity user = userRepository.findById(userId).get();
+
+            if(user.getId().equals(dealComment.getUser().getId())){
+                if(dealComment.getUpIdx() != 0){
+                    deal.setComment(deal.getComment() - 1);
+                    dealRepository.save(deal);
+
+                    dealCommentRepository.delete(dealComment);
+
+                    Optional<NoticeEntity> notice = noticeRepository.findByNoticeTypeAndFromUserIdAndPostTypeAndPostIdxAndCommentIdx("reply", user.getId(), "deal", deal.getIdx(), dealComment.getIdx());
+                    if(notice.isPresent()){
+                        noticeRepository.delete(notice.get());
+                    }
+                }
+            }else{
+                List<DealCommentEntity> replyCommentList = dealCommentRepository.findByUpIdx(idx);
+                int size = replyCommentList.size();
+
+                if(!replyCommentList.isEmpty()){
+                    deal.setComment(deal.getComment() - size - 1);
+                    dealRepository.save(deal);
+
+                    dealCommentRepository.deleteAllInBatch(replyCommentList);
+
+                    List<NoticeEntity> noticeList = noticeRepository.findAllByNoticeTypeAndFromUserIdAndPostTypeAndPostIdxAndCommentUpIdx("reply", user.getId(), "deal", deal.getIdx(), dealComment.getIdx());
+
+                    if(!noticeList.isEmpty()){
+                        noticeRepository.deleteAllInBatch(noticeList);
+                    }
+                    dealCommentRepository.delete(dealComment);
+
+                    Optional<NoticeEntity> notice = noticeRepository.findByNoticeTypeAndFromUserIdAndPostTypeAndPostIdxAndCommentIdx("comment", user.getId(), "deal", deal.getIdx(), dealComment.getIdx());
+
+                    if(notice.isPresent()){
+                        noticeRepository.delete(notice.get());
+                    }
+                }
+            }
             return true;
         } else{
             return false;
@@ -180,6 +243,11 @@ public class DealServiceImpl implements DealService {
                 userLikeDealsRepository.delete(userLikeDeals);
                 deal.setLikes(deal.getLikes() - 1);
                 dealRepository.save(deal);
+
+                Optional<NoticeEntity> notice = noticeRepository.findByNoticeTypeAndFromUserIdAndPostTypeAndPostIdx("like",user.getId(),"deal",deal.getIdx());
+                if(notice.isPresent()){
+                    noticeRepository.delete(notice.get());
+                }
             } else{
                 UserLikeDealsEntity userLikeDeals = UserLikeDealsEntity
                         .builder()
@@ -189,6 +257,17 @@ public class DealServiceImpl implements DealService {
                 userLikeDealsRepository.save(userLikeDeals);
                 deal.setLikes(deal.getLikes() + 1);
                 dealRepository.save(deal);
+
+                NoticeEntity notice = NoticeEntity.builder()
+                        .noticeType("like")
+                        .user(deal.getUser())
+                        .fromUserId(user.getId())
+                        .postType("deal")
+                        .postIdx(deal.getIdx())
+                        .time(userLikeDeals.getTime())
+                        .build();
+
+                noticeRepository.save(notice);
             }
             return true;
         } else{
